requirements.txt
python-telegram-bot==20.7
Pillow==10.0.0
python-dotenv==1.0.0
nest_asyncio==1.5.8 